+++ app/build.gradle.kts
@@ -43,6 +43,10 @@ android {
 
 dependencies {
     implementation ("com.google.android.material:material:1.8.0")
+    implementation ("androidx.navigation:navigation-fragment-ktx:2.5.3")
+    implementation ("androidx.navigation:navigation-ui-ktx:2.5.3")
+    implementation ("androidx.fragment:fragment-ktx:1.5.6")
+    implementation ("com.google.android.material:material:1.8.0")
     implementation ("androidx.viewpager2:viewpager2:1.0.0")
     implementation ("androidx.fragment:fragment-ktx:1.5.5")
     implementation ("io.insert-koin:koin-android:3.3.0")
