+++ app/src/main/AndroidManifest.xml
@@ -15,26 +15,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Material3.DayNight.NoActionBar"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.media_library.activity.MediaLibraryActivity"
-            android:exported="false" />
 
         <activity
-            android:name=".ui.audio_player.activity.AudioPlayerActivity"
-            android:exported="false"/>
-
-        <activity
-            android:name=".ui.search.activity.SearchActivity"
-            android:windowSoftInputMode="stateVisible|adjustResize"
-            android:exported="false" />
-
-        <activity
-            android:name=".ui.settings.activity.SettingsActivity"
-            android:exported="false"/>
-
-        <activity
-            android:name=".ui.main.activity.MainActivity"
-            android:exported="true">
+            android:name=".ui.root.RootActivity"
+            android:exported="true"
+            android:windowSoftInputMode="adjustPan|stateHidden">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
