+++ app/src/main/java/com/example/playlistmaker/data/media_library/impl/MediaLibraryRepositoryImpl.kt
@@ -3,8 +3,8 @@ package com.example.playlistmaker.data.media_library.impl
 import com.example.playlistmaker.R
 import com.example.playlistmaker.domain.media_library.MediaLibraryRepository
 import com.example.playlistmaker.domain.models.TabData
-import com.example.playlistmaker.ui.media_library.fragments.favourite_fragment.FavouriteFragment
-import com.example.playlistmaker.ui.media_library.fragments.playlist_fragment.PlaylistFragment
+import com.example.playlistmaker.ui.favourite.fragment.FavouriteFragment
+import com.example.playlistmaker.ui.playlist.fragment.PlaylistFragment
 
 class MediaLibraryRepositoryImpl : MediaLibraryRepository {
     override fun getTabData() = listOf(
