+++ app/src/main/java/com/example/playlistmaker/di/DataModule.kt
@@ -6,7 +6,6 @@ import com.example.playlistmaker.app.TrackMapper
 import com.example.playlistmaker.data.NetworkClient
 import com.example.playlistmaker.data.local.SearchHistoryStorage
 import com.example.playlistmaker.data.local.SearchHistoryStorageImpl
-import com.example.playlistmaker.data.main.impl.MainExternalNavigatorImpl
 import com.example.playlistmaker.data.network.ItunesApiService
 import com.example.playlistmaker.data.network.RetrofitNetworkClient
 import com.example.playlistmaker.data.network.SearchHistoryRepositoryImpl
@@ -14,7 +13,6 @@ import com.example.playlistmaker.data.search.SearchRepositoryImpl
 import com.example.playlistmaker.data.settings.impl.SettingsRepositoryImpl
 import com.example.playlistmaker.data.sharing.impl.ExternalNavigatorImpl
 import com.example.playlistmaker.domain.api.SearchHistoryRepository
-import com.example.playlistmaker.domain.main.MainExternalNavigator
 import com.example.playlistmaker.domain.search.SearchRepository
 import com.example.playlistmaker.domain.settings.SettingsRepository
 import com.example.playlistmaker.domain.sharing.ExternalNavigator
@@ -53,10 +51,6 @@ val dataModule = module {
         SearchHistoryStorageImpl(get(), get())
     }
 
-
-    single<MainExternalNavigator> {
-        MainExternalNavigatorImpl(androidContext())
-    }
 
     single<ExternalNavigator> {
         ExternalNavigatorImpl(androidContext())
