+++ app/src/main/java/com/example/playlistmaker/di/InteractorModule.kt
@@ -2,8 +2,6 @@ package com.example.playlistmaker.di
 
 import com.example.playlistmaker.domain.api.SearchHistoryInteractor
 import com.example.playlistmaker.domain.impl.SearchHistoryInteractorImpl
-import com.example.playlistmaker.domain.main.MainInteractor
-import com.example.playlistmaker.domain.main.impl.MainInteractorImpl
 import com.example.playlistmaker.domain.media_library.MediaLibraryInteractor
 import com.example.playlistmaker.domain.player.AudioPlayerInteractor
 import com.example.playlistmaker.domain.search.SearchInteractor
@@ -25,12 +23,6 @@ val interactorModule = module {
     single<SearchHistoryInteractor> {
         SearchHistoryInteractorImpl(
             repository = get()
-        )
-    }
-
-    single<MainInteractor> {
-        MainInteractorImpl(
-            mainExternalNavigator = get()
         )
     }
