+++ app/src/main/java/com/example/playlistmaker/di/ViewModelModule.kt
@@ -1,9 +1,8 @@
 package com.example.playlistmaker.di
 
 import com.example.playlistmaker.ui.audio_player.view_model.AudioPlayerViewModel
-import com.example.playlistmaker.ui.main.view_model.MainViewModel
-import com.example.playlistmaker.ui.media_library.fragments.favourite_fragment.FavouriteViewModel
-import com.example.playlistmaker.ui.media_library.fragments.playlist_fragment.PlaylistViewModel
+import com.example.playlistmaker.ui.favourite.view_model.FavouriteViewModel
+import com.example.playlistmaker.ui.playlist.view_model.PlaylistViewModel
 import com.example.playlistmaker.ui.media_library.view_model.MediaLibraryViewModel
 import com.example.playlistmaker.ui.search.view_model.SearchViewModel
 import com.example.playlistmaker.ui.settings.view_model.SettingsViewModel
@@ -19,11 +18,6 @@ val viewModelModule = module {
         )
     }
 
-    viewModel {
-        MainViewModel(
-            mainInteractor = get()
-        )
-    }
 
     viewModel {
         SearchViewModel(
