--- app/src/main/java/com/example/playlistmaker/domain/main/MainExternalNavigator.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain.main
-
-
-interface MainExternalNavigator {
-    fun openSearch()
-    fun openMediaLibrary()
-    fun openSettings()
-}
