--- app/src/main/java/com/example/playlistmaker/domain/main/MainInteractor.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.main
-
-interface MainInteractor {
-    fun searchButton()
-    fun mediaLibraryButton()
-    fun settingsButton()
-}
