--- app/src/main/java/com/example/playlistmaker/domain/main/impl/MainInteractorImpl.kt
@@ -1,21 +1,0 @@
-package com.example.playlistmaker.domain.main.impl
-
-import com.example.playlistmaker.domain.main.MainExternalNavigator
-import com.example.playlistmaker.domain.main.MainInteractor
-
-class MainInteractorImpl(
-    private val mainExternalNavigator: MainExternalNavigator,
-) : MainInteractor {
-
-    override fun searchButton() {
-        mainExternalNavigator.openSearch()
-    }
-
-    override fun mediaLibraryButton() {
-        mainExternalNavigator.openMediaLibrary()
-    }
-
-    override fun settingsButton() {
-        mainExternalNavigator.openSettings()
-    }
-}
