+++ app/src/main/java/com/example/playlistmaker/ui/favourite/view_model/FavouriteViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.ui.favourite.view_model
+
+import androidx.lifecycle.ViewModel
+
+class FavouriteViewModel : ViewModel(){
+}
