--- app/src/main/java/com/example/playlistmaker/ui/media_library/fragments/favourite_fragment/FavouriteFragment.kt
@@ -1,32 +1,0 @@
-package com.example.playlistmaker.ui.media_library.fragments.favourite_fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.playlistmaker.databinding.FragmentFavouriteTracksBinding
-import org.koin.androidx.viewmodel.ext.android.viewModel
-
-class FavouriteFragment : Fragment() {
-    private lateinit var binding: FragmentFavouriteTracksBinding
-    private val viewModel by viewModel<FavouriteViewModel>()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentFavouriteTracksBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.favouriteTracksEmptyPlaceholder.visibility = View.VISIBLE
-    }
-
-    companion object {
-        fun newInstance() = FavouriteFragment()
-    }
-}
