--- app/src/main/java/com/example/playlistmaker/ui/media_library/fragments/favourite_fragment/FavouriteViewModel.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.ui.media_library.fragments.favourite_fragment
-
-import androidx.lifecycle.ViewModel
-
-class FavouriteViewModel : ViewModel(){
-}
