--- app/src/main/java/com/example/playlistmaker/ui/media_library/fragments/playlist_fragment/PlaylistFragment.kt
@@ -1,33 +1,0 @@
-package com.example.playlistmaker.ui.media_library.fragments.playlist_fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.playlistmaker.databinding.FragmentPlaylistBinding
-import org.koin.androidx.viewmodel.ext.android.viewModel
-
-class PlaylistFragment : Fragment() {
-    private lateinit var binding: FragmentPlaylistBinding
-    private val viewModel by viewModel<PlaylistViewModel>()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentPlaylistBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.playlistsEmptyPlaceholder.visibility = View.VISIBLE
-        binding.createPlaylistButton.visibility = View.VISIBLE
-    }
-
-    companion object {
-        fun newInstance() = PlaylistFragment()
-    }
-}
