--- app/src/main/java/com/example/playlistmaker/ui/media_library/fragments/playlist_fragment/PlaylistViewModel.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.ui.media_library.fragments.playlist_fragment
-
-import androidx.lifecycle.ViewModel
-
-class PlaylistViewModel : ViewModel() {
-}
