+++ app/src/main/java/com/example/playlistmaker/ui/playlist/view_model/PlaylistViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.ui.playlist.view_model
+
+import androidx.lifecycle.ViewModel
+
+class PlaylistViewModel : ViewModel() {
+}
