+++ app/src/main/java/com/example/playlistmaker/ui/root/RootActivity.kt
@@ -1,0 +1,40 @@
+package com.example.playlistmaker.ui.root
+
+import android.os.Bundle
+import android.view.View
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.commit
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.ActivityRootBinding
+import com.example.playlistmaker.ui.search.fragment.SearchFragment
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+class RootActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityRootBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityRootBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.container_view) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        binding.bottomNavigationView.setupWithNavController(navController)
+
+        navController.addOnDestinationChangedListener { _, destination, _ ->
+            when(destination.id) {
+                R.id.searchFragment, R.id.mediaLibraryFragment, R.id.settingsFragment -> {
+                    binding.bottomNavigationView.visibility = View.VISIBLE
+                }
+                else -> {
+                    binding.bottomNavigationView.visibility = View.GONE
+                }
+            }
+        }
+    }
+}
