+++ app/src/main/res/color/bottom_nav_icon_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/selected_color" android:state_checked="true" />
+    <item android:color="@color/unselected_color" android:state_checked="false" />
+</selector>
