+++ app/src/main/res/drawable/ic_search_light.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M16.579,14.95H15.721L15.417,14.656C16.481,13.418 17.122,11.81 17.122,10.061C17.122,6.161 13.961,3 10.061,3C6.161,3 3,6.161 3,10.061C3,13.961 6.161,17.122 10.061,17.122C11.81,17.122 13.418,16.481 14.656,15.417L14.95,15.721V16.579L20.381,22L22,20.381L16.579,14.95ZM10.061,14.95C7.356,14.95 5.173,12.766 5.173,10.061C5.173,7.356 7.356,5.173 10.061,5.173C12.766,5.173 14.95,7.356 14.95,10.061C14.95,12.766 12.766,14.95 10.061,14.95Z"
+      android:fillColor="#1A1B22"/>
+</vector>
