--- app/src/main/res/layout/activity_main.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    style="@style/MainBackgroundStyle"
-    android:padding="16dp"
-    tools:ignore="ExtraText">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text= "@string/title"
-        style="@style/MainTextViewStyle"
-        android:textSize="22sp"
-        android:fontFamily="@font/ys_display_medium"
-        android:textFontWeight="500"/>
-
-    <Button
-        android:id="@+id/search_button"
-        style="@style/MainButtonStyle"
-        android:layout_marginTop="40dp"
-        android:text="@string/search"
-        app:icon="@drawable/search" />
-
-    <Button
-        android:id="@+id/media_library_button"
-        style="@style/MainButtonStyle"
-        android:layout_marginTop="16dp"
-        android:text="@string/media_library"
-        app:icon="@drawable/media_library" />
-
-    <Button
-        android:id="@+id/settings_button"
-        style="@style/MainButtonStyle"
-        android:layout_marginTop="16dp"
-        android:text="@string/settings"
-        android:layout_marginBottom="28dp"
-        app:icon="@drawable/settings" />
-
-    </LinearLayout>
