--- app/src/main/res/layout/activity_media_library.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/white"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/header_height"
-        android:orientation="horizontal"
-        style="@style/SettingBackgroundStyle"
-        android:gravity="center_vertical">
-
-        <ImageView
-            android:id="@+id/back_button"
-            android:layout_width="@dimen/size_second"
-            android:layout_height="@dimen/size_second"
-            android:layout_marginTop="@dimen/margin_start_third"
-            android:layout_marginBottom="@dimen/margin_start_third"
-            android:layout_marginStart="@dimen/margin_start_third"
-            android:src="@drawable/back1"
-            style="@style/BackArrowStyle" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/size_second"
-            android:text="@string/media_library"
-            style="@style/SettingHeaderStyle"
-            android:textSize="@dimen/text_size_large"
-            android:paddingStart="@dimen/margin_start"
-            android:fontFamily="@font/ys_display_medium"
-            android:textFontWeight="500"
-            android:gravity="start|center"
-            android:layout_marginStart="0dp" />
-
-    </LinearLayout>
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/tabLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="@dimen/text_size1"
-        android:fontFamily="@font/ys_display_medium"
-        android:textFontWeight="500"
-        app:tabMode="fixed"
-        app:tabIndicatorHeight="2dp"
-        app:tabIndicatorFullWidth="true"
-        app:tabRippleColor="@android:color/transparent"
-        style="@style/TabIndicator"/>
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/viewPager"
-        style="@style/FragmentBg"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-</LinearLayout>
