--- app/src/main/res/layout/activity_settings.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical"
-    style="@style/SettingBackgroundStyle">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/header_height"
-        android:orientation="horizontal"
-        style="@style/SettingBackgroundStyle"
-        android:gravity="center_vertical">
-
-        <ImageView
-            android:id="@+id/back_button"
-            android:layout_width="@dimen/size_second"
-            android:layout_height="@dimen/size_second"
-            android:layout_marginTop="@dimen/margin_start_third"
-            android:layout_marginBottom="@dimen/margin_start_third"
-            android:layout_marginStart="@dimen/margin_start_third"
-            android:src="@drawable/back1"
-            style="@style/BackArrowStyle" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="@string/settings"
-            style="@style/SettingHeaderStyle"
-            android:textSize="@dimen/text_size_large"
-            android:paddingStart="12dp"
-            android:fontFamily="@font/ys_display_medium"
-            android:textFontWeight="500"
-            android:gravity="start|center"
-            android:layout_marginStart="0dp"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        style="@style/SettingButtonStyle"
-        android:layout_marginTop="@dimen/margin_top">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:gravity="start|center_vertical"
-            android:padding="@dimen/padding_standard"
-            android:text="@string/dark_theme"
-            style="@style/DarkThemeButtonStyle"
-            android:textSize="@dimen/text_size_medium"
-            android:fontFamily="@font/ys_display_regular"
-            android:textFontWeight="400"
-            android:layout_weight="1"
-            android:width="0dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <com.google.android.material.switchmaterial.SwitchMaterial
-            android:id="@+id/themeSwitcher"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:checked="false"
-            android:paddingEnd="@dimen/margin_start"
-            style="@style/SwitchButtonStyle"
-            tools:ignore="RtlSymmetry,UseSwitchCompatOrMaterialXml" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/share_app_button"
-        style="@style/SettingButtonStyle"
-        android:gravity="center_vertical"
-        android:padding="@dimen/padding_standard"
-        android:clickable="true"
-        android:focusable="true">
-
-        <TextView
-            style="@style/SettingTextViewStyle"
-            android:text="@string/share_app"/>
-
-        <ImageView
-            style="@style/SettingImageViewStyle"
-            android:src="@drawable/share" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/write_to_support_button"
-        style="@style/SettingButtonStyle"
-        android:gravity="center_vertical"
-        android:padding="@dimen/padding_standard"
-        android:clickable="true"
-        android:focusable="true">
-
-        <TextView
-            style="@style/SettingTextViewStyle"
-            android:text="@string/write_to_support"/>
-
-        <ImageView
-            style="@style/SettingImageViewStyle"
-            android:src="@drawable/support" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/user_agreement_button"
-        style="@style/SettingButtonStyle"
-        android:gravity="center_vertical"
-        android:padding="@dimen/padding_standard"
-        android:clickable="true"
-        android:focusable="true" >
-
-        <TextView
-            style="@style/SettingTextViewStyle"
-            android:text="@string/user_agreement" />
-
-        <ImageView
-            style="@style/SettingImageViewStyle"
-            android:src="@drawable/agreement" />
-    </LinearLayout>
-
-</LinearLayout>
