+++ app/src/main/res/layout/fragment_favourite_tracks.xml
@@ -1,7 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.example.playlistmaker.ui.favourite.fragment.FavouriteFragment">
 
     <ImageView
         android:id="@+id/favouriteTracksEmptyPlaceholder"
