+++ app/src/main/res/layout/fragment_media_library.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:background="@color/white"
+    android:orientation="vertical"
+    tools:context="com.example.playlistmaker.ui.media_library.fragment.MediaLibraryFragment">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/header_height"
+        android:orientation="horizontal"
+        style="@style/SettingBackgroundStyle"
+        android:gravity="center_vertical">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="@dimen/size_second"
+            android:text="@string/media_library"
+            style="@style/SettingHeaderStyle"
+            android:textSize="@dimen/text_size_large"
+            android:paddingStart="@dimen/padding_standard"
+            android:fontFamily="@font/ys_display_medium"
+            android:textFontWeight="500"
+            android:gravity="start|center"
+            android:layout_marginStart="0dp" />
+
+    </LinearLayout>
+
+    <com.google.android.material.tabs.TabLayout
+        android:id="@+id/tabLayout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:textSize="@dimen/text_size1"
+        android:fontFamily="@font/ys_display_medium"
+        android:textFontWeight="500"
+        app:tabMode="fixed"
+        app:tabIndicatorHeight="2dp"
+        app:tabIndicatorFullWidth="true"
+        app:tabRippleColor="@android:color/transparent"
+        style="@style/TabIndicator"/>
+
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/viewPager"
+        style="@style/FragmentBg"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
+</LinearLayout>
