+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/ic_search_light"
+        android:title="@string/search" />
+
+    <item
+        android:id="@+id/mediaLibraryFragment"
+        android:icon="@drawable/ic_media_light"
+        android:title="@string/media_library" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/ic_settings_light"
+        android:title="@string/settings" />
+</menu>
