+++ app/src/main/res/values-night/bottom_nav_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="selected_color">#3772E7</color>
+    <color name="unselected_color">#FFFFFF</color>
+</resources>
