+++ app/src/main/res/values-night/darkthemestyles.xml
@@ -129,4 +129,17 @@
         <item name="android:background">@color/black_for_settings</item>
     </style>
 
+    <style name="IconSearchStyle">
+        <item name="android:icon">@drawable/ic_search_dark</item>
+    </style>
+
+    <style name="IconMediaStyle">
+        <item name="android:icon">@drawable/ic_media_dark</item>
+    </style>
+
+    <style name="IconSettingsStyle">
+        <item name="android:icon">@drawable/ic_settings_dark</item>
+    </style>
+
+
 </resources>
